[RUBY-2802] Refactor CertificatesController into WasteCarriersEngine namespace and update views, routes, and assets accordingly.
