+++ config/initializers/assets.rb
@@ -12,4 +12,4 @@ Rails.application.config.assets.version = "1.0"
 # application.js, application.css, and all non-JS/CSS in app/assets folder are already added.
 Rails.application.config.assets.precompile += %w[*.jpg *.png]
 Rails.application.config.assets.precompile += %w[receipt_email.js]
-Rails.application.config.assets.precompile += %w[certificates.css]
+Rails.application.config.assets.precompile += %w[certificates.css environment_agency_logo.png]
