+++ config/locales/certificates.en.yml
@@ -1,43 +1,44 @@
 en:
-  certificates:
-    errors:
-      token: "You do not have permission to view this page"
-    show:
-      heading: "Certificate of Registration under the Waste (England and Wales) Regulations 2011"
-      authority: "Regulation authority"
-      name: "Name"
-      nccc_address_pretext: "Address"
-      nccc_address_line_1: "National Customer Contact Centre"
-      nccc_address_line_2: "99 Parkway Avenue"
-      nccc_address_line_3: "Sheffield"
-      nccc_address_line_4: "S9 4WF"
-      telephone_number_pretext: "Telephone number"
-      nccc_number: "03708 506506"
-      certify_message: "The Environment Agency certify that the following information is entered in the register which they maintain under regulation 28 of the Waste (England and Wales) Regulations 2011."
-      carrier_table_heading: "Carriers details"
-      entity_display_name: "Name of registered carrier"
-      business_name_if_applicable: "Business name (if applicable)"
-      partners: "Partners"
-      registered: "Registered as"
-      registration_number: "Registration number"
-      business_address: "Address of place of business"
-      carrier_telephone_number: "Telephone number"
-      registration_date: "Date of registration"
-      expiry_date: "Expiry date of registration (unless revoked)"
-      cert_created_date: "This certificate was created on %{date}. These details are correct at the time of certificate generation."
-      cert_copy_text: "This copy has been issued under Regulation 6 of Waste (England and Wales) Amendment Act 2014 by the Environment Agency. This is copy number %{certificate_version} of the certificate."
-      heading2: "Making changes to your registration"
-      lower_renewal: "Your registration will last indefinitely so does not need to be renewed but you must update your registration details if they change, within 28 days of the change."
-      upper_renewal: "Your registration will last %{expires_after_pluralized} and will need to be renewed after this period. If any of your details change, you must notify us within 28 days of the change."
-      year: "year"
-      update_phone: "You can do this by calling the Environment Agency."
-      registered_as:
-        lower: "A lower tier waste carrier, broker and dealer"
-        upper: "An upper tier waste %{registration_type}"
-      carrier_dealer: "carrier and dealer"
-      broker_dealer: "broker and dealer"
-      carrier_broker_dealer: "carrier, broker and dealer"
-    confirm_email:
-      legend: "Confirm your email address"
-    process_email:
-      error: "Invalid email address"
+  waste_carriers_engine:
+    certificates:
+      errors:
+        token: "You do not have permission to view this page"
+      show:
+        heading: "Certificate of Registration under the Waste (England and Wales) Regulations 2011"
+        authority: "Regulation authority"
+        name: "Name"
+        nccc_address_pretext: "Address"
+        nccc_address_line_1: "National Customer Contact Centre"
+        nccc_address_line_2: "99 Parkway Avenue"
+        nccc_address_line_3: "Sheffield"
+        nccc_address_line_4: "S9 4WF"
+        telephone_number_pretext: "Telephone number"
+        nccc_number: "03708 506506"
+        certify_message: "The Environment Agency certify that the following information is entered in the register which they maintain under regulation 28 of the Waste (England and Wales) Regulations 2011."
+        carrier_table_heading: "Carriers details"
+        entity_display_name: "Name of registered carrier"
+        business_name_if_applicable: "Business name (if applicable)"
+        partners: "Partners"
+        registered: "Registered as"
+        registration_number: "Registration number"
+        business_address: "Address of place of business"
+        carrier_telephone_number: "Telephone number"
+        registration_date: "Date of registration"
+        expiry_date: "Expiry date of registration (unless revoked)"
+        cert_created_date: "This certificate was created on %{date}. These details are correct at the time of certificate generation."
+        cert_copy_text: "This copy has been issued under Regulation 6 of Waste (England and Wales) Amendment Act 2014 by the Environment Agency. This is copy number %{certificate_version} of the certificate."
+        heading2: "Making changes to your registration"
+        lower_renewal: "Your registration will last indefinitely so does not need to be renewed but you must update your registration details if they change, within 28 days of the change."
+        upper_renewal: "Your registration will last %{expires_after_pluralized} and will need to be renewed after this period. If any of your details change, you must notify us within 28 days of the change."
+        year: "year"
+        update_phone: "You can do this by calling the Environment Agency."
+        registered_as:
+          lower: "A lower tier waste carrier, broker and dealer"
+          upper: "An upper tier waste %{registration_type}"
+        carrier_dealer: "carrier and dealer"
+        broker_dealer: "broker and dealer"
+        carrier_broker_dealer: "carrier, broker and dealer"
+      confirm_email:
+        legend: "Confirm your email address"
+      process_email:
+        error: "Invalid email address"
