+++ config/routes.rb
@@ -340,30 +340,28 @@ WasteCarriersEngine::Engine.routes.draw do
     get "/back", to: "forms#go_back", as: "go_back_forms"
   end
 
-  namespace :registrations, param: :reg_identifier, path: "/bo/registrations" do
-    get "certificate", to: "certificates#show"
-    get "pdf_certificate", to: "certificates#pdf"
-    get "certificate_confirm_email", to: "certificates#confirm_email"
-    post "certificate_process_email", to: "certificates#process_email"
-  end
+  get ":reg_identifier/certificate", to: "certificates#show", as: "certificate"
+  get ":reg_identifier/pdf_certificate", to: "certificates#pdf", as: "pdf_certificate"
+  get ":reg_identifier/certificate_confirm_email", to: "certificates#confirm_email", as: "certificate_confirm_email"
+  post ":reg_identifier/certificate_process_email", to: "certificates#process_email", as: "certificate_process_email"
 
   mount DefraRubyEmail::Engine => "/email"
 
   # See http://patrickperey.com/railscast-053-handling-exceptions/
   get "(errors)/:status",
-    to: "errors#show",
-    constraints: { status: /\d{3}/ },
-    as: "error"
+      to: "errors#show",
+      constraints: { status: /\d{3}/ },
+      as: "error"
 
   # Renew via magic link token
   get "/renew/:token",
-    to: "renews#new",
-    as: "renew"
+      to: "renews#new",
+      as: "renew"
 
   # Deregister via magic link token
   get "/deregister/:token",
-    to: "deregisters#new",
-    as: "deregister"
+      to: "deregisters#new",
+      as: "deregister"
 
   # Static pages with HighVoltage
   resources :pages, only: [:show], controller: "pages"
