+++ sonar-project.properties
@@ -27,3 +27,4 @@ sonar.sourceEncoding=UTF-8
 
 sonar.ruby.coverage.reportPaths=coverage/coverage.json
 sonar.ruby.rubocop.reportPaths=rubocop-result.json
+sonar.exclusions=app/views/waste_carriers_engine/certificates/show.html.erb
